// tests/enums/slugs.ts
export enum singleSlugs {
  eGamesPage = 'egame',
  liveSlotPage = 'live-slot',
  liveCasinoPage = 'live-casino',
  pagePage = 'page',
  promoPage = 'promotion',
  articlePage = 'article',
  bannerPage = 'banner'
}

export enum pluralSlugs {
  eGamesPage = 'egames',
  liveSlotPage = 'live-slots',
  liveCasinoPage = 'live-casinos',
  pagePage = 'pages',
  promoPage = 'promotions',
  articlePage = 'articles',
  bannersPage = 'banners'
}